backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = "AbuseBox FastAPI backend"
authors = ["bekkaze <bekkaze7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
sqlalchemy = "^2.0.36"
pydantic = "^2.10.3"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
